# docs/swagger.yaml
basePath: /
definitions:
  internal_api.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      passkey_id:
        type: string
    type: object
  internal_api.TelegramAuthRequest:
    properties:
      auth_date:
        type: integer
      first_name:
        type: string
      hash:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      username:
        type: string
    type: object
  internal_api.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Backend for personal finances
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Flizix API
  version: "1.0"
paths:
  /auth/telegram:
    post:
      consumes:
      - application/json
      description: Validate Telegram hash, search or create user emitting a JWT o
        a cookie.
      parameters:
      - description: telegram data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_api.TelegramAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.UserResponse'
      summary: Login con Telegram
      tags:
      - auth
  /health:
    get:
      description: Returns simple message to validate server is running
      produces:
      - text/plain
      responses:
        "200":
          description: Flizix service OK
          schema:
            type: string
      summary: Check health system
      tags:
      - system
  /me:
    get:
      description: Get userdata using JWT on cookie.
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.UserResponse'
        "401":
          description: Not authorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get profile
      tags:
      - users
  /users:
    post:
      consumes:
      - application/json
      description: Register new user on DB with name, email and passkey ID.
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_api.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api.UserResponse'
        "400":
          description: Invalid JSON
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create new user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"